Move files around for clarity
