+++ frontend/src/AuthenticatedApp/paperScripts/ScriptViewSubpage/index.tsx
