+++ frontend/src/AuthenticatedApp/paperScripts/ScriptsPage/ScriptsTableRow.tsx
