+++ frontend/src/AuthenticatedApp/paperSetup/index.ts
@@ -1,0 +1,11 @@
+import SetupPage from "./SetupPage";
+import QuestionAllocationSubpage from "./QuestionAllocationSubpage";
+import ScriptMappingSubpage from "./ScriptMappingSubpage";
+import ScriptTemplateSubpage from "./ScriptTemplateView";
+
+export {
+  QuestionAllocationSubpage,
+  ScriptMappingSubpage,
+  ScriptTemplateSubpage
+};
+export default SetupPage;
